requirements.txt
accelerate==0.31.0
boto3==1.37.33
clean_fid==0.1.35
datasets==2.20.0
deepspeed==0.15.1
diffusers==0.35.0
duckdb==1.1.2
einops==0.8.0
fire==0.5.0
lightning==2.4.0
numpy==1.26.4
omegaconf==2.3.0
Pillow==9.5.0
sentencepiece==0.2.0
torch==2.5.0
torchvision==0.20.0
transformers==4.43.3
wandb==0.16.6
webdataset==0.2.86
zstandard==0.23.0
hf-transfer==0.1.9